% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rearrange}
\alias{.rearrange}
\title{k' is the number of variables in the model
fit' is the fit function of the regression model
names' is a vector of variable names in the order you used}
\usage{
.rearrange(k, fit, names)
}
\description{
k' is the number of variables in the model
fit' is the fit function of the regression model
names' is a vector of variable names in the order you used
}
