% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicSEM-package.R
\docType{package}
\name{GenomicSEM-package}
\alias{GenomicSEM}
\alias{GenomicSEM-package}
\title{GenomicSEM: Structural Equation Modeling Based on GWAS Summary Statistics}
\description{
Create, run, and report Genomic Structural Equation Models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GenomicSEM/GenomicSEM}
  \item Report bugs at \url{https://github.com/GenomicSEM/GenomicSEM/issues}
}

}
\author{
\strong{Maintainer}: Andrew Grotzinger \email{agrotzin@utexas.edu}

Authors:
\itemize{
  \item Andrew Grotzinger \email{agrotzin@utexas.edu}
}

Other contributors:
\itemize{
  \item Matthijs van der Zee [contributor]
  \item Mijke Rhemtulla [contributor]
  \item Hill Ip [contributor]
  \item Michel Nivard [contributor]
  \item Elliot Tucker-Drob [contributor]
}

}
\keyword{internal}
