<<<<<<< HEAD
\name{ldsc}
\alias{ldsc}
\title{build a covariance structure using LD score regression in R}
\description{
Function to run LD score regression (\url{https://github.com/bulik/ldsc}) to compute the genetic covariance between a
series of traits based on genome wide summary statistics obtained from GWAS. The results generated by this function are
necessary and sufficient to fit structural equation models (or other multivariate models).
}
\usage{
ldsc(traits,sample.prev,population.prev,ld,wld,trait.names=NULL,sep_weights=FALSE,chr=22,n.blocks=200,log.file='ldsc.log' \dots)

}
\arguments{
  \item{traits}{
    A vector of strings pointing to LDSC munged files for the traits to include in a GenomicSEM model.
  }
  \item{sample.prev}{
    A vector of sample prevalences for dichotomous traits and \code{NA} for continuous traits.
  }
  \item{population.prev}{
    A vector of population prevalences for dichotomous traits or \code{NA} for continuous traits.
  }
  \item{ld}{
    A string containing the path to the folder in which the LD scores used in the analysis are located.
    Expects LD scores formatted as required by the original LD score regression software.
  }
  \item{wld}{
    A string containing the path to the folder in which the LD score weights used in the analysis are located.
    Expects LD scores formatted as required by the original LD score regression software.
  }
  \item{trait.names}{
    A character vector specifying how the traits should be named in the genetic covariance (S) matrix. These variable
    names can subsequently be used in later steps for model specification. If no value is provided, the function will
    automatically name the variables using the generic from of V1-VX.
  }
  \item{sep_weights}{
    Logical indicating wheter the weights are different from the LD scores used for the regression, defaults to FALSE.
  }
  \item{chr}{
    Number of chromosomes over which the LDSC weights are split, defalts to 22 (Human) but can be changed for other
    species.
  }
  \item{n.blocks}{
    Number of blocks to use in the jackknife procedure deployed to estimate entries in V: higher values will be optimal
    if you have a large number of variables but also slower; defaults to 200.
  }
}

\value{
  The function returns a list with the following named entries (in this order):
  \item{V}{
    variance covariance matrix of the parameter estimates in S
  }
  \item{S}{
    estimated genetic covariance matrix
  }
  \item{I}{
    matrix containing the "inflation" intercepts due to GWAS sample substructures (diagonal terms) or cross-GWAS samples
    overlaps (in terms of subjects, off-diagonal)
  }
  \item{N}{
    a vector containing the sample size (for the genetic variances) and the geometric mean sample sizes (sqrt(N1,N2))
    between two samples for the covariances
  }
  \item{m}{
    number of SNPs used to compute the LD scores with.
  }
  \item{V_stand}{
    correlation matrix of the parameter estimates in S
  }
  \item{S_stand}{
    estimated genetic correlation matrix
  }
  \item{pS}{
    LD-score regression coefficients p-values
  }
  \item{pI}{
    matrix containing T-test p-values for "inflation" intercepts different from 1 (diagonal) and 1 (off-diagonal)
  }
  \item{n}{
    number of blocks used in the jackknife procedure
  }
}


\examples{

}

